manage

+++ app/Http/Controllers/ManageController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\CategoryModel;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -9,23 +10,35 @@ use Image;
 
 class ManageController extends Controller
 {
+    public function index()
+    {
+        return View('manage.index');
+    }
+
     public function profile()
     {
         $data = Auth::User();
-        return View('manage.profile',compact('data'));
+        return View('manage.profile', compact('data'));
+    }
+
+    public function add()
+    {
+        $category = CategoryModel::all();
+        return View('manage.add',compact('category'));
     }
 
     public function updateprofile(Request $request)
     {
         $data = User::find(Auth::User()->id);
         $data->page_name = $request->page_name;
+        $data->link_name = $request->link_name;
         $data->bio = $request->bio;
         $data->border_c_1 = $request->border_c_1;
         $data->border_c_2 = $request->border_c_2;
         $data->border_c_3 = $request->border_c_3;
         $image = $request->file('image');
         if ($image) {
-            $name_gen = "logo_profile_".$data->link_name;
+            $name_gen = "logo_profile_" . $data->link_name;
             // $img_ext = strtolower($image->getClientOriginalExtension());
             $img_name = $name_gen . '.' . 'webp';
             $upload_location = 'images/profile/';
@@ -40,7 +53,7 @@ class ManageController extends Controller
         }
         $image_cover = $request->file('image_cover');
         if ($image_cover) {
-            $name_gen = "cover_image_".$data->link_name;
+            $name_gen = "cover_image_" . $data->link_name;
             // $img_ext = strtolower($image_cover->getClientOriginalExtension());
             $img_name = $name_gen . '.' . 'webp';
             $upload_location = 'images/banner/';
@@ -48,7 +61,9 @@ class ManageController extends Controller
             $data->image_cover = $full_path;
             $old_image_cover = $request->old_image_cover;
             if (file_exists($old_image_cover)) {
-                unlink($old_image_cover);
+                if($request->old_image_cover !== 'images/banner/default.webp'){
+                    unlink($old_image_cover);
+                }
             }
             $img = Image::make($image_cover->path());
             $img->encode('webp', 90)->resize(1500, 1500, function ($constraint) {
